refactor: Move error codes to constants.py
