+++ README.md
@@ -62,13 +62,13 @@ And depending on which error code range you've opted into, it will tell you
 
 ## Error codes
 
-| Code   | Description                                         |
-|--------|-----------------------------------------------------|
-| TC001 | Move import into a type-checking block  |
-| TC002 | Move third-party import into a type-checking block |
-| TC003 | Found multiple type checking blocks |
+| Code   | Description                                                                        |
+|--------|------------------------------------------------------------------------------------|
+| TC001 | Move application import into a type-checking block                                 |
+| TC002 | Move third-party import into a type-checking block                                 |
+| TC003 | Found multiple type checking blocks                                                |
 | TC004 | Move import out of type-checking block. Import is used for more than type hinting. |
-| TC005 | Empty type-checking block |
+| TC005 | Empty type-checking block                                                          |
 
 ## Choosing how to handle forward references
