+++ tests/test_pydantic.py
@@ -9,7 +9,7 @@ import textwrap
 
 import pytest
 
-from flake8_type_checking.codes import TC002
+from flake8_type_checking.constants import TC002
 from tests import _get_error
