+++ tests/test_tc001.py
@@ -13,7 +13,7 @@ import textwrap
 
 import pytest
 
-from flake8_type_checking.codes import TC001
+from flake8_type_checking.constants import TC001
 from tests import _get_error, mod
 
 examples = [
