+++ tests/test_tc100.py
@@ -10,7 +10,7 @@ One thing to note: futures imports should always be at the top of a file, so we 
 
 import pytest
 
-from flake8_type_checking.codes import TC100
+from flake8_type_checking.constants import TC100
 from tests import _get_error, mod
 
 examples = [
