+++ tests/test_tc101.py
@@ -6,7 +6,7 @@ import textwrap
 
 import pytest
 
-from flake8_type_checking.codes import TC101
+from flake8_type_checking.constants import TC101
 from tests import _get_error
 
 examples = [
