+++ tests/test_tc200.py
@@ -6,7 +6,7 @@ import textwrap
 
 import pytest
 
-from flake8_type_checking.codes import TC200
+from flake8_type_checking.constants import TC200
 from tests import _get_error
 
 examples = [
