+++ tests/test_tc201.py
@@ -6,7 +6,7 @@ import textwrap
 
 import pytest
 
-from flake8_type_checking.codes import TC201
+from flake8_type_checking.constants import TC201
 from tests import _get_error
 
 examples = [
